Make maxLength and party name options

+++ src/main/scala/mot/message/ClientHello.scala
@@ -4,13 +4,13 @@ import mot.buffer.ReadBuffer
 import mot.buffer.WriteBuffer
 import java.nio.charset.StandardCharsets
 
-case class ClientHello(protocolVersion: Byte, sender: String, maxLength: Short) extends MessageBase {
+case class ClientHello(protocolVersion: Byte, sender: String, maxLength: Int) extends MessageBase {
   
   def writeToBuffer(writeBuffer: WriteBuffer) = {
     writeBuffer.put(MessageType.ClientHello.id.toByte)
     writeBuffer.put(protocolVersion)
     MessageBase.writeByteSizeByteField(writeBuffer, sender.getBytes(StandardCharsets.US_ASCII))
-    writeBuffer.putShort(maxLength)
+    writeBuffer.putInt(maxLength)
   }
   
   override def toString() = s"ClientHello(protocolVersion=$protocolVersion,sender=$sender,maxLength=$maxLength)"
@@ -22,7 +22,7 @@ object ClientHello {
   def factory(readBuffer: ReadBuffer, maxLength: Int) = {
     val version = readBuffer.get()
     val sender = new String(MessageBase.readByteSizeByteField(readBuffer), StandardCharsets.US_ASCII)
-    val maxLength = readBuffer.getShort()
+    val maxLength = readBuffer.getInt()
     ClientHello(version, sender, maxLength)
   }
