+++ src/main/scala/mot/message/ServerHello.scala
@@ -4,13 +4,13 @@ import mot.buffer.ReadBuffer
 import mot.buffer.WriteBuffer
 import java.nio.charset.StandardCharsets
 
-case class ServerHello(protocolVersion: Byte, name: String, maxLength: Short) extends MessageBase {
+case class ServerHello(protocolVersion: Byte, name: String, maxLength: Int) extends MessageBase {
 
   def writeToBuffer(writeBuffer: WriteBuffer) = {
     writeBuffer.put(MessageType.ServerHello.id.toByte)
     writeBuffer.put(protocolVersion)
     MessageBase.writeByteSizeByteField(writeBuffer, name.getBytes(StandardCharsets.US_ASCII))
-    writeBuffer.putShort(maxLength)
+    writeBuffer.putInt(maxLength)
   }
   
   override def toString() = s"ServerHello(protocolVersion=$protocolVersion,maxLength=$maxLength)"
@@ -22,7 +22,7 @@ object ServerHello{
   def factory(readBuffer: ReadBuffer, maxLength: Int) = {
     val version = readBuffer.get()
     val server = new String(MessageBase.readByteSizeByteField(readBuffer), StandardCharsets.US_ASCII)
-    val maxLength = readBuffer.getShort()
+    val maxLength = readBuffer.getInt()
     ServerHello(version, server, maxLength)
   }
